requirements.txt
openai-whisper
torch
requests
numpy
pandas
scikit-learn
joblib
